moved tests/unit/lbrynet_daemon/test_docs.py into tests/unit/test_cli.py
